\docType{package}
\name{factRdata}
\alias{factRdata}
\title{Sample custom transcriptomes to supplement factR}
\description{
    This package contain custom transcriptomes generated from 3
    generations of RNA-seq experiments (bulk, single-cell, long-read) in the
    form of GTF files. These files will serve as inputs to the vignette
    generated in the package factR.
}
\keyword{package}
